+++ .prettierrc
@@ -1,7 +1,7 @@
 {
-  "semi": false,
+  "semi": true,
   "singleQuote": true,
   "tabWidth": 2,
-  "useTabs": false,
+  "useTabs": true,
   "trailingComma": "es5"
 }
