+++ src/context/TimeContext/index.js
@@ -1,0 +1,2 @@
+export { TimeContextProvider } from './TimeContext';
+export { useTimeContext } from './useTimeContext';
